feat(connect): Full rewite of all LIRMM variants using ::connect method
