--- modules/rsdf/panda2_lirmm/Panda2LIRMM.rsdf
@@ -1,4 +1,0 @@
-<?xml version="1.0" ?>
-<robot name="panda_default">
-
-</robot>
